+++ spa-routing-core/src/main/kotlin/com/sparouting/contract/SpaRouteFactories.kt
@@ -3,9 +3,10 @@ package com.sparouting.contract
 fun route(
   path: String,
   id: String,
-  parameters: List<SpaRouteParameter> = emptyList()
+  parameters: List<SpaRouteParameter> = emptyList(),
+  queryParameters: List<SpaRouteParameter> = emptyList()
 ): SpaRouteDefinition {
-  return SpaRouteDefinition(path, id, parameters)
+  return SpaRouteDefinition(path, id, parameters, queryParameters)
 }
 
 fun string(name: String): SpaRouteParameter {