+++ spa-routing-core/src/main/kotlin/com/sparouting/contract/SpaTypedRoute.kt
@@ -4,15 +4,15 @@ open class SpaTypedRoute(
   override val applicationId: String,
   override val routeId: String
 ) : SpaRouteKey {
-  operator fun invoke(): SpaRouteTarget {
-    return target()
-  }
-
-  protected fun target(parameters: Map<String, String> = emptyMap()): SpaRouteTarget {
+  protected fun target(
+    parameters: Map<String, String> = emptyMap(),
+    queryParameters: Map<String, List<String>> = emptyMap()
+  ): SpaRouteTarget {
     return SpaRouteTarget(
       applicationId = applicationId,
       routeId = routeId,
-      parameters = parameters
+      parameters = parameters,
+      queryParameters = queryParameters
     )
   }
 }