+++ spa-routing-spring-boot-autoconfigure/src/main/kotlin/io/github/caseymcguire/sparouting/spring/web/SpaRouterFunctionFactory.kt
@@ -40,6 +40,10 @@ class SpaRouterFunctionFactory(
       return ServerResponse.status(properties.server.invalidPathParameterStatus).build()
     }
 
+    if (!route.hasValidQueryParameterValues(request.params())) {
+      return ServerResponse.status(properties.server.invalidQueryParameterStatus).build()
+    }
+
     val response = routeResponseEvaluator.evaluate(
       applicationRules = config.rules,
       routeRules = config.getRouteRules(route),